requirements.txt
requests
beautifulsoup4
trafilatura
pandas
tqdm
python-dotenv
lxml
feedparser
opendatasets
